pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "linkedin-cli"
version = "0.1.0"
description = "A terminal-first CLI for LinkedIn feed, profiles, search, and core social actions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "jlebrero" }]
keywords = ["linkedin", "cli", "social", "feed", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]
dependencies = [
    "beautifulsoup4>=4.12",
    "browser-cookie3>=0.20",
    "click>=8.1",
    "httpx>=0.27",
    "linkedin-api>=2.2.0",
    "lxml>=5.3",
    "playwright>=1.52",
    "PyYAML>=6.0",
    "requests>=2.32",
    "rich>=13.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-mock>=3.14",
    "ruff>=0.11",
]

[project.urls]
Homepage = "https://github.com/frizynn/linkedin-cli"
Repository = "https://github.com/frizynn/linkedin-cli"
Issues = "https://github.com/frizynn/linkedin-cli/issues"
Documentation = "https://github.com/frizynn/linkedin-cli#readme"
Changelog = "https://github.com/frizynn/linkedin-cli/blob/main/CHANGELOG.md"
Security = "https://github.com/frizynn/linkedin-cli/security"

[project.scripts]
linkedin = "linkedin_cli.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["linkedin_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
